Add AXL_XFER_BEST to autodetect the best transfer API
Add a new AXL_XFER_BEST type for AXL_Create().  This allows the caller
to leave it up to AXL to choose the best API for the particular
node.  For example, if you're running on an IBM node, use the BB API.
If you're running on a Cray, use DataWarp.  Otherwise use sync.  That
way the caller doesn't have to have hardware-specific transfers in their
checkpointing code.