Don't put axl_pthread_data pointer in KVTree
This patch uses a linked list to lookup which axl_pthread_data is associated
with each AXL ID.  Why put it in a linked list instead of just storing the
pdata in the kvtree?  Because if we put it in a kvtree, and the app dies, the
pdata pointer becomes stale, and would get erroneously freed as part of an
AXL_Stop().  Instead we use axl_pthread_data_lookup(), axl_pthread_data_add(),
and axl_pthread_data_remove() to access the data.  This makes it so the pdata
is ephemeral, only existing while the app is running.

Signed-off-by: Tony Hutter <hutter2@llnl.gov>