BBAPI fallback to pthreads if FS doesn't support extents
If the user requests a BBAPI transfer, and one or more of the source
or destination paths in the file list aren't to a filesystem that
BBAPI supports, then fallback to doing a pthreads transfer for all the
files in the file list. This gets around BBAPI giving you a "file not
found" when it tries to transfer to/from a filesystem that doesn't
support extents (like NFS).

Fixes: #61