Use temporary file name when transferring
This patch will add an temporary AXL extension to the file while
it is copying.  The extension is removed after the file copy
is complete.  For example, if you're copying to /tmp/file1,
AXL will actually copy to /tmp/file1._AXL, and then rename to
/tmp/file1 at the end of the transfer.

Additionally, a BB API transfer will also encode the transfer
handle number into the extension like:

/tmp/file1._AXL13456

That way we can later recover the transfer handle number.

Fixes: #66