Add AXL_XFER_STATE_FILE xfer type
Add AXL_XFER_STATE_FILE transfer type.  This type is used to tell
AXL that you want to use the transfer type specified in the
state_file.  This is useful if you're trying to resume a transfer
from a previous state_file, since it saves you from having to
remember the old transfer type you used.