Check destination file sizes before resuming transfers
This changes AXL_Dispatch() to first check the destination file
sizes before resuming the copy.  That way, if the files were already
fully transferred at the time of resume, and the source files are
gone (as would be the case with a SCR post-stage), then simply
finalize the destination files (remove their ._AXL extension,
apply metadata, etc).