Use 'bbcmd' to check BB transfer status in post-stage
We want to be able to check the BBAPI transfer status while running from the
job node.  This is needed for when we do a BBAPI resume transfer from
the 2nd post-stage script on IBM machines.  In that case we don't have access
to the BBAPI, but we can spawn off 'bbcmd' to see if transfer we wanted
to resume had already successfully transferred.  This patch adds in the
'bbcmd' spawn and check for those cases.