Fix BBAPI tag generation
The BBAPI requires us to pass  a 64-bit tag that will never be
repeated on a node.  Previously, we used the unix timestamp +
thread ID for the tag.  This lead to duplicate tags when SCR would start
two checkpoints in the same second.  This patch adds in an additional
2-bit quarter second field to the tag along with a quarter second wait, to
ensure no identical tags are ever generated on the node.